Updated auth role routes
