+++ views/dashListener.ejs
@@ -23,7 +23,7 @@
     </nav>
 
     <div class="big-container">
-        <form class="form-a" action="/dashMusician" method="POST">
+        <form class="form-a" action="/dashListener" method="POST">
             <input
             type="text"
             id="playlist_name"
@@ -38,7 +38,7 @@
             </div>
         </form>
 
-        <form class="form-a" action="/dashMusician" method="POST">
+        <form class="form-a" action="/dashListener" method="POST">
             <input
             type="text"
             id="playlist_name2"
@@ -53,7 +53,7 @@
             </div>
         </form>
 
-        <form class="form-a" action="/dashMusician" method="POST">
+        <form class="form-a" action="/dashListener" method="POST">
             <div>
                 <input
                 type="text"
@@ -78,7 +78,7 @@
             </div>
         </form>
 
-        <form class="form-a" action="/dashMusician" method="POST">
+        <form class="form-a" action="/dashListener" method="POST">
             <div>
                 <input
                 type="text"
@@ -109,7 +109,7 @@
             <div class="big-container">
             <table class="tablecontainer"> 
             <th class="playlist-header">
-                <form action="/dashboard" method="POST">
+                <form action="/dashListener" method="POST">
                     <input
                     type="submit"
                     value="<%= playlist[i].playlist_name %>"
